refactor: 把枚举 enum ythtbbs_user_login_type 及其对应的方法转移到 libbmy 中
